style: rename code
