+++ example/lib/src/curve_app/app/app_page.dart
