+++ example/lib/src/curve_app/web/web_page.dart
