+++ example/lib/src/curve_app/widgets/ball_animation.dart
