+++ example/lib/src/curve_app/widgets/code_preview/bounce_code_preview.dart
