+++ example/lib/src/curve_app/widgets/code_preview/gravity_code_preview.dart
