+++ example/lib/src/curve_app/widgets/curve_illustration.dart
