+++ example/lib/src/curve_app/widgets/curve_option/bounce_option.dart
