+++ example/lib/src/curve_app/widgets/curve_option/cubic_option.dart
