+++ example/lib/src/curve_app/widgets/curve_option/gravity_option.dart
