+++ example/lib/src/curve_app/widgets/effect_animation.dart
